0.1.2a

+++ src/main/java/austeretony/oxygen_exchange/common/network/server/SPSendExchangeRequest.java
@@ -17,11 +17,11 @@ public class SPSendExchangeRequest extends ProxyPacket {
 
     @Override
     public void write(PacketBuffer buffer, INetHandler netHandler) {
-        buffer.writeShort(this.index);
+        buffer.writeInt(this.index);
     }
 
     @Override
     public void read(PacketBuffer buffer, INetHandler netHandler) {
-        ExchangeManagerServer.instance().sendExchangeRequest(getEntityPlayerMP(netHandler), buffer.readShort());
+        ExchangeManagerServer.instance().sendExchangeRequest(getEntityPlayerMP(netHandler), buffer.readInt());
     }
 }
